# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "license": {}
    },
    "paths": {
        "/binddevice": {
            "post": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "绑定设备",
                "parameters": [
                    {
                        "description": "UserDevice",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dao.BindDeviceDao"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/control": {
            "post": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "远程控制",
                "parameters": [
                    {
                        "description": "message",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Message"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getdevice": {
            "get": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "获取局域网设备",
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/getstatus": {
            "get": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Control"
                ],
                "summary": "获取状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "openId",
                        "name": "openId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "获取JSON",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "小程序登录",
                "parameters": [
                    {
                        "description": "res",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Res"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\": -1, \"message\": \"抱歉未找到相关信息\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dao.BindDeviceDao": {
            "type": "object",
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "model.Message": {
            "type": "object",
            "required": [
                "deviceId",
                "type"
            ],
            "properties": {
                "deviceId": {
                    "type": "string"
                },
                "during": {
                    "description": "持续时间",
                    "type": "integer"
                },
                "type": {
                    "description": "1打开 2 关闭",
                    "type": "integer"
                }
            }
        },
        "model.Res": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "error": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                }
            }
        }
    }
}